backend/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-resized==1.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
exceptiongroup==1.2.0
gunicorn==21.2.0
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.5
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==7.4.4
pytest-django==4.7.0
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13